python-mqtt-client/requirements.txt
asttokens==2.4.1
colorama==0.4.6
docopt==0.6.1
executing==2.1.0
Flask==0.10.1
gitdb==4.0.11
GitPython==0.3.2rc1
icecream==2.1.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mypi==0.0.3
paho-mqtt==2.1.0
Pygments==2.18.0
semantic-version==2.3.0
six==1.16.0
smmap==5.0.1
Werkzeug==3.0.3
